+++ src/ctrl/VectorDataManager.js
@@ -2530,7 +2530,13 @@ class VectorDataManager extends HistorizedManager {
 
 	_isLanduse(feature) {
 		const t = feature.properties.tags;
-		return t.landuse || ["school","university","college","hospital"].includes(t.amenity) || t.boundary || t["disused:boundary"];
+		return (
+			t.landuse
+			|| ["school","university","college","hospital"].includes(t.amenity)
+			|| t.boundary
+			|| t["disused:boundary"]
+			|| (t.leisure === "sports_centre" && !t.building)
+		);
 	}
 
 	/**